+++ trixnity-messenger/src/commonMain/kotlin/de/connect2x/trixnity/messenger/DI.kt
@@ -1,11 +1,6 @@
 package de.connect2x.trixnity.messenger
 
-import de.connect2x.trixnity.client.MatrixClientConfiguration
 import de.connect2x.trixnity.client.ModuleFactory
-import de.connect2x.trixnity.clientserverapi.client.MatrixClientServerApiClientFactory
-import de.connect2x.trixnity.core.MSC3814
-import de.connect2x.trixnity.core.serialization.events.EventContentSerializerMappings
-import de.connect2x.trixnity.core.serialization.events.default
 import de.connect2x.trixnity.messenger.export.TimelineEventContentToString
 import de.connect2x.trixnity.messenger.export.TimelineEventContentToStringImpl
 import de.connect2x.trixnity.messenger.export.exportModule
@@ -31,7 +26,6 @@ import de.connect2x.trixnity.messenger.util.LeaveRoom
 import de.connect2x.trixnity.messenger.util.LeaveRoomImpl
 import de.connect2x.trixnity.messenger.util.MatrixMarkdownFlavour
 import de.connect2x.trixnity.messenger.util.MatrixMarkdownFlavourImpl
-import de.connect2x.trixnity.messenger.util.RelevantTimelineEvents
 import de.connect2x.trixnity.messenger.util.Search
 import de.connect2x.trixnity.messenger.util.SearchImpl
 import de.connect2x.trixnity.messenger.util.SharedDataHandler
@@ -127,6 +121,7 @@ import de.connect2x.trixnity.messenger.viewmodel.roomlist.RoomListElementViewMod
 import de.connect2x.trixnity.messenger.viewmodel.roomlist.RoomListViewModelFactory
 import de.connect2x.trixnity.messenger.viewmodel.roomlist.SearchGroupViewModelFactory
 import de.connect2x.trixnity.messenger.viewmodel.settings.AccountSetupViewModelFactory
+import de.connect2x.trixnity.messenger.viewmodel.settings.AccountSingleViewModelFactory
 import de.connect2x.trixnity.messenger.viewmodel.settings.AccountsViewModelFactory
 import de.connect2x.trixnity.messenger.viewmodel.settings.AppInfoViewModelFactory
 import de.connect2x.trixnity.messenger.viewmodel.settings.AppearanceSettingsViewModelFactory
@@ -138,7 +133,6 @@ import de.connect2x.trixnity.messenger.viewmodel.settings.NotificationSettingsAl
 import de.connect2x.trixnity.messenger.viewmodel.settings.NotificationSettingsSingleAccountViewModelFactory
 import de.connect2x.trixnity.messenger.viewmodel.settings.PrivacySettingsAllAccountsViewModelFactory
 import de.connect2x.trixnity.messenger.viewmodel.settings.PrivacySettingsSingleAccountViewModelFactory
-import de.connect2x.trixnity.messenger.viewmodel.settings.ProfileSingleViewModelFactory
 import de.connect2x.trixnity.messenger.viewmodel.settings.ProfilesSettingsSingleViewModelFactory
 import de.connect2x.trixnity.messenger.viewmodel.settings.ProfilesSettingsViewModelFactory
 import de.connect2x.trixnity.messenger.viewmodel.settings.UserSettingsViewModelFactory
@@ -197,51 +191,21 @@ import org.koin.dsl.bind
 import org.koin.dsl.module
 import kotlin.time.Clock
 
-fun interface ConfigureMatrixClientConfiguration {
-    operator fun MatrixClientConfiguration.invoke()
-}
-
-fun interface DebugName {
-    operator fun invoke(): String
-}
-
 fun createTrixnityMessengerDefaultModuleFactories(): List<ModuleFactory> = listOf(
     {
         module {
             single<Clock> { Clock.System }
             single<TimeZone> { TimeZone.currentSystemDefault() }
-
-            single<ConfigureMatrixClientConfiguration>(named("DefaultConfigureMatrixClientConfiguration")) {
-                val config = get<MatrixMessengerConfiguration>()
-                val relevantTimelineEvents = get<RelevantTimelineEvents>()
-                val eventContentSerializerMappings = getAll<EventContentSerializerMappings>()
-                ConfigureMatrixClientConfiguration {
-                    name = getOrNull<DebugName>()?.invoke()
-                    markOwnMessageAsRead = true
-                    enableExternalNotifications = true
-                    httpClientEngine = config.httpClientEngine
-                    httpClientConfig = config.httpClientConfig
-                    @OptIn(MSC3814::class)
-                    experimentalFeatures.enableMSC3814 = true
-                    lastRelevantEventFilter =
-                        { relevantTimelineEvents.isRelevantTimelineEvent(it.content) }
-                    if (eventContentSerializerMappings.isNotEmpty()) {
-                        modulesFactories += {
-                            module {
-                                single<EventContentSerializerMappings> {
-                                    eventContentSerializerMappings
-                                        .fold(EventContentSerializerMappings.default) { a, b -> a + b }
-                                }
-                            }
-                        }
-                    }
-                    getOrNull<MatrixClientServerApiClientFactory>()?.let {
-                        matrixClientServerApiClientFactory = it
-                    }
-                }
+            single<MatrixClientFactory> {
+                MatrixClientFactoryImpl(
+                    secretByteArrays = get(),
+                    createRepositoriesModule = get(),
+                    createMediaStoreModule = get(),
+                    createCryptoDriverModule = get(),
+                    appCoroutineContext = get<CoroutineScope>().coroutineContext,
+                    messengerConfiguration = get(),
+                )
             }
-
-            single<MatrixClientFactory> { MatrixClientFactory }
             single<MatrixClients> {
                 MatrixClientsImpl(
                     matrixClientFactory = get(),
@@ -249,12 +213,7 @@ fun createTrixnityMessengerDefaultModuleFactories(): List<ModuleFactory> = listO
                     settings = get(),
                     config = get(),
                     secretByteArrays = get(),
-                    createRepositoriesModule = get(),
-                    createMediaStoreModule = get(),
-                    createCryptoDriverModule = get(),
-                    appCoroutineContext = get<CoroutineScope>().coroutineContext,
                     i18n = get(),
-                    configurer = getAll()
                 )
             }.apply {
                 bind<AutoCloseable>()
@@ -264,7 +223,6 @@ fun createTrixnityMessengerDefaultModuleFactories(): List<ModuleFactory> = listO
             single<TimelineEventContentToString> { TimelineEventContentToStringImpl(get()) }
             single<Initials> { InitialsImpl(get()) }
             single<VerifyAccount> { VerifyAccountImpl() }
-            single<RelevantTimelineEvents> { RelevantTimelineEvents }
 
             single<Languages> { DefaultLanguages }
             single<I18n> { I18n(get(), get(), get(), get()) }
@@ -384,7 +342,7 @@ private fun settingsViewModels() = module {
     single<AccountsViewModelFactory> { AccountsViewModelFactory }
     single<ProfilesSettingsViewModelFactory> { ProfilesSettingsViewModelFactory }
     single<ProfilesSettingsSingleViewModelFactory> { ProfilesSettingsSingleViewModelFactory }
-    single<ProfileSingleViewModelFactory> { ProfileSingleViewModelFactory }
+    single<AccountSingleViewModelFactory> { AccountSingleViewModelFactory }
     single<UserSettingsViewModelFactory> { UserSettingsViewModelFactory }
     single<PrivacySettingsAllAccountsViewModelFactory> { PrivacySettingsAllAccountsViewModelFactory }
     single<PrivacySettingsSingleAccountViewModelFactory> { PrivacySettingsSingleAccountViewModelFactory }