+++ trixnity-messenger/src/commonMain/kotlin/de/connect2x/trixnity/messenger/MatrixMessengerConfiguration.kt
@@ -1,5 +1,6 @@
 package de.connect2x.trixnity.messenger
 
+import de.connect2x.trixnity.client.MatrixClientConfiguration
 import de.connect2x.trixnity.client.ModuleFactory
 import de.connect2x.trixnity.messenger.util.mb
 import io.ktor.client.HttpClient
@@ -111,6 +112,11 @@ data class MatrixMessengerConfiguration(
     var cryptoDriver: CryptoDriver = CryptoDriver.VODOZEMAC,
 
     /**
+     * Consider using [clientConfiguration], as it can be called multiple times.
+     */
+    var client: MatrixClientConfiguration.() -> Unit = { },
+
+    /**
      * Specify a [HttpClientEngine]. It is highly recommended to set it and share it within an application.
      */
     override var httpClientEngine: HttpClientEngine? = null,
@@ -132,6 +138,15 @@ data class MatrixMessengerConfiguration(
      */
     var modulesFactories: List<ModuleFactory> = createTrixnityMessengerDefaultModuleFactories(),
 ) : MatrixMessengerBaseConfiguration {
+
+    fun clientConfiguration(config: MatrixClientConfiguration.() -> Unit) {
+        val original = client
+        client = {
+            original()
+            config()
+        }
+    }
+
     /**
      * A list of feature toggles for the Matrix Messenger.
      */