+++ CHANGELOG.md
@@ -17,10 +17,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - UI,ANDROID,IOS,MACOS: First-class support for notification count app icon badges
 - SDK: A separate package which provides Nunito as Font for the application
 - SDK, UI: Toggle between bundled and system Font
+- SDK: Introduce `MatrixMessengerConfiguration.clientConfiguration` to allow for customizing the `MatrixClient` via
+  configuration
 
 ### Changed
 
-- SDK: Update SysNotify to 2.6.1
+- DEPENDENCY: Update SysNotify to 2.6.1
+- DEPENDENCY: Update Trixnity to 5.5.0
 - SDK: Use builtin Gradle distribution for dockerized CI jobs
 - SDK: Update README
 - SDK: smoke tests for Android with more logging
@@ -28,8 +31,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - CI: Use new Android emulator docker image
 - SDK: Don't ship font per default
 - UI: Group buttons in pdf viewer layout
+- SDK: **breaking change**: `AccountSingleViewModel` now has `cancelEditDisplayName()` and `saveDisplayName()` instead
+  of the surrounding `AccountsViewModel`; also introduced `AccountSingleViewModelFactory` which correctly gets a
+  `MatrixClientViewModelContext`
+- UI: **breaking change**: introduced `AccountSingleSettingsView` to allow customizations of settings for an account
+  instead of overriding the settings container for all accounts
+- SDK: **breaking change**: `MatrixClientServerApiClientFactory` cannot be used inside Trixnity Messenger DI anymore.
+  Extend `MatrixClientFactory` instead
+- SDK: **breaking change**: `EventContentSerializerMappings` passed to the DI are not picked up anymore. Instead,
+  `CustomEventContentSerializerMappings` are now picked up from the
+  `MatrixMessengerConfiguration.clientConfiguration.moduleFactories`
 
 ### Deprecated
+
+- SDK: `ProfileSingleViewModelFactory` -> use `AccountSingleViewModelFactory`
 
 ### Removed